requirements.txt
# Updated on 2025-06-23

python-dotenv==1.1.0

bs4==0.0.2
chromadb==1.0.13

langgraph==0.4.8
langchain==0.3.26
langchain-community==0.3.26
langchain-google-genai==2.1.5
langchain-tavily==0.2.4
